+++ package.json
@@ -51,29 +51,29 @@
     "react-native-animatable": "1.4.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.5.5",
-    "@babel/runtime": "^7.6.3",
+    "@babel/core": "^7.20.5",
+    "@babel/runtime": "7.22.15",
     "@react-native-community/eslint-config": "^3.2.0",
     "@semantic-release/git": "^9.0.0",
     "@types/jest": "^29.5.6",
     "@types/react": "^18.2.33",
     "@types/react-native": "^0.72.5",
     "@typescript-eslint/eslint-plugin": "^6.9.0",
-    "@typescript-eslint/parser": "^2.4.0",
+    "@typescript-eslint/parser": "^5.4.0",
     "babel-core": "7.0.0-bridge.0",
     "babel-jest": "^29.7.0",
-    "babel-plugin-module-resolver": "3.1.3",
+    "babel-plugin-module-resolver": "^4.1.0",
     "detox": "^20.13.3",
     "eslint": "^8.52.0",
     "eslint-plugin-prettier": "^5.0.1",
-    "husky": "^3.0.9",
+    "husky": "^8.0.3",
     "jest": "^29.7.0",
     "lint-staged": "^15.0.2",
     "metro-react-native-babel-preset": "^0.77.0",
     "postinstall": "^0.5.1",
     "prettier": "^3.0.3",
     "react": "^18.2.0",
-    "react-native": "^0.72.6",
+    "react-native": "0.72.3",
     "react-test-renderer": "^18.2.0",
     "semantic-release": "17.3.0",
     "typescript": "^5.2.2"
