+++ src/types.ts
@@ -1,5 +1,5 @@
-import {Animation, CustomAnimation} from 'react-native-animatable';
-import {NativeSyntheticEvent, NativeTouchEvent} from 'react-native';
+import { Animation, CustomAnimation } from 'react-native-animatable';
+import { NativeSyntheticEvent, NativeTouchEvent } from 'react-native';
 
 export type OrNull<T> = null | T;
 
@@ -8,6 +8,16 @@ export type Animations = {
   animationIn: string;
   animationOut: string;
 };
+
+export type CustomAnimationType =
+  | 'slideInDown'
+  | 'slideInUp'
+  | 'slideInLeft'
+  | 'slideInRight'
+  | 'slideOutDown'
+  | 'slideOutUp'
+  | 'slideOutLeft'
+  | 'slideOutRight'
 
 export type Orientation =
   | 'portrait'
@@ -28,4 +38,4 @@ export type OnOrientationChange = (
 ) => void;
 
 export interface GestureResponderEvent
-  extends NativeSyntheticEvent<NativeTouchEvent> {}
+  extends NativeSyntheticEvent<NativeTouchEvent> { }
